# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/address/:address_id": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete an address by its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete one address.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The address id",
                        "name": "address_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update an address by its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update an address.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The address id",
                        "name": "address_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an address",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateAddressInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Commerce_Address"
                        }
                    }
                }
            }
        },
        "/api/user/address/create": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new address user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create an address.",
                "parameters": [
                    {
                        "description": "the body to create a new address",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateAddressInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Commerce_Address"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateAddressInput": {
            "type": "object",
            "properties": {
                "address_city": {
                    "type": "string"
                },
                "address_country": {
                    "type": "string"
                },
                "address_latitude": {
                    "type": "string"
                },
                "address_longitude": {
                    "type": "string"
                },
                "address_postcode": {
                    "type": "integer"
                },
                "address_province": {
                    "type": "string"
                },
                "address_street": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UpdateAddressInput": {
            "type": "object",
            "properties": {
                "address_city": {
                    "type": "string"
                },
                "address_country": {
                    "type": "string"
                },
                "address_latitude": {
                    "type": "string"
                },
                "address_longitude": {
                    "type": "string"
                },
                "address_postcode": {
                    "type": "integer"
                },
                "address_province": {
                    "type": "string"
                },
                "address_street": {
                    "type": "string"
                }
            }
        },
...
        "models.Commerce_Address": {
            "type": "object",
            "properties": {
                "address_city": {
                    "type": "string"
                },
                "address_country": {
                    "type": "string"
                },
                "address_id": {
                    "type": "integer"
                },
                "address_latitude": {
                    "type": "string"
                },
                "address_longitude": {
                    "type": "string"
                },
                "address_postcode": {
                    "type": "integer"
                },
                "address_province": {
                    "type": "string"
                },
                "address_street": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}